Move authorization component
