+++ src/applications/disability-benefits/686/sass/686-dependent-status.scss
@@ -6,6 +6,7 @@
 @import "~@department-of-veterans-affairs/formation/sass/modules/m-modal";
 @import "~@department-of-veterans-affairs/formation/sass/modules/m-omb-info";
 @import "../../../../platform/forms/sass/m-form-confirmation";
+@import "../../../../platform/forms/sass/_m-applications";
 
 .dependent-title {
     padding-bottom: 0;
