+++ src/platform/forms/sass/_m-applications.scss
@@ -2,3 +2,9 @@
   margin-bottom: 1em;
   margin-top: 0;
 }
+
+.sip-authorization-container {
+  .sip-authorization-message {
+      margin: 0 0 1.5em;
+  }
+}
