+++ src/platform/forms/tests/components/AuthorizationComponent.unit.spec.jsx
